Crate: add font weight keyword parsing
